1.因 id long 类型会在 js 中超长，若使用默认json输出方式会丢失精度，此处暂降所有 long 型强转换为string输出 2.菜单管理 新增、修改、删除

+++ src/main/java/com/ssy/petition/common/CommonResult.java
@@ -5,7 +5,7 @@ package com.ssy.petition.common;
  */
 public class CommonResult<T> {
 
-    private long code;
+    private int code;
 
     private String message;
 
@@ -14,7 +14,7 @@ public class CommonResult<T> {
     protected CommonResult() {
     }
 
-    protected CommonResult(long code, String message, T data) {
+    protected CommonResult(int code, String message, T data) {
         this.code = code;
         this.message = message;
         this.data = data;
@@ -94,11 +94,11 @@ public class CommonResult<T> {
         return new CommonResult<T>(ResultCode.FORBIDDEN.getCode(), ResultCode.FORBIDDEN.getMessage(), data);
     }
 
-    public long getCode() {
+    public int getCode() {
         return code;
     }
 
-    public void setCode(long code) {
+    public void setCode(int code) {
         this.code = code;
     }
